Add version check logic comparing local and remote Moulin version
This change introduces automatic version checking for the Moulin.
When Moulin is executed directly (not via ninja), it fetches the remote
"setup.py" from GitHub and compares its version with the locally
installed one.

If the remote version is newer, a warning is printed with an offer
on to upgrade Moulin.

Signed-off-by: Mykhailo Androsiuk <mykhailo_androsiuk@epam.com>