yocto: layers not being removed from bblayers.conf
Fix touches the yocto_add_layers ninja rule.
It checks the given set of layers against the ones previously used
in the build.
If a cache file exists, it compares the provided layers with
the ones supplied by "bitbake-layers show-layers".
The "path" column is from "bitbake-layers show-layers" and converted
into an array.
Then the array is canonicalized to bring all paths to a single,
standard format.
After that, the paths are converted to relative paths.
The provided layer set is also converted into an array.
To speed things up, the logic first compares the number of elements
in both arrays.
If the count is the same, it compares the actual contents
of the arrays.
If any of these checks detect a difference, the existing layers
are removed and replaced with the new set, and the hash is updated.
If the cache file does not exist, the logic adds the layers to Yocto
and saves the cache for future use.

Signed-off-by: Mykhailo Androsiuk <mykhailo_androsiuk@epam.com>